# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AppFormat": {
        "enum": [
          "onepanel",
          "baota",
          "flame"
        ],
        "type": "string"
      },
      "AppMetadata": {
        "properties": {
          "architectures": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "category": {
            "type": "string"
          },
          "default_version": {
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/AppFormat"
          },
          "key": {
            "type": "string"
          },
          "logo": {
            "type": [
              "string",
              "null"
            ]
          },
          "min_memory_mb": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "modes": {
            "items": {
              "$ref": "#/components/schemas/InstallMode"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "readme": {
            "type": [
              "string",
              "null"
            ]
          },
          "recommended": {
            "description": "商店首页推荐位标记（v0.7.0）",
            "type": "boolean"
          },
          "short_desc_en": {
            "type": [
              "string",
              "null"
            ]
          },
          "short_desc_zh": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "versions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "key",
          "name",
          "category",
          "short_desc_zh",
          "tags",
          "format",
          "modes",
          "versions",
          "default_version",
          "architectures"
        ],
        "type": "object"
      },
      "AppStoreListResponse": {
        "properties": {
          "packages": {
            "items": {
              "$ref": "#/components/schemas/AppMetadata"
            },
            "type": "array"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "packages",
          "total"
        ],
        "type": "object"
      },
...
      "InstallMode": {
        "enum": [
          "container",
          "native",
          "wasm"
        ],
        "type": "string"
      },
      "InstalledApp": {
        "properties": {
          "access_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "container_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "install_path": {
            "type": "string"
          },
          "launch_count": {
            "description": "启动次数（常用应用排序，v0.7.0）",
            "format": "int64",
            "type": "integer"
          },
          "mode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "package_key": {
            "type": "string"
          },
          "params_json": {
            "type": "string"
          },
          "port": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "package_key",
          "name",
          "version",
          "mode",
          "status",
          "install_path",
          "params_json",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "must_change_password": {
            "description": "是否需强制修改密码（新装面板首次登录）",
            "type": "boolean"
          },
          "refresh_token": {
            "description": "Refresh Token（长过期，默认 24 小时；用于换取新 Access Token）",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "token": {
            "description": "Access Token（短过期，默认 15 分钟）",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "refresh_token",
          "username",
          "role",
          "must_change_password"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FlamePanel 面板后端 REST API（OpenAPI 3.x，由 utoipa 编译期生成）",
    "license": {
      "name": ""
    },
    "title": "FlamePanel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/app-store/installed": {
      "get": {
        "operationId": "list_installed",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstalledApp"
                  },
                  "type": "array"
                }
              }
            },
            "description": "已安装应用列表"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "已安装应用列表",
        "tags": [
          "app_store"
        ]
      }
    },
    "/api/app-store/packages": {
      "get": {
        "operationId": "list_packages",
        "parameters": [
          {
            "description": "分类过滤",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppStoreListResponse"
                }
              }
            },
            "description": "包列表"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "应用商店包列表",
        "tags": [
          "app_store"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "登录成功"
          },
          "401": {
            "description": "凭据错误"
          },
          "429": {
            "description": "尝试过于频繁"
          }
        },
        "summary": "登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "刷新成功"
          },
          "401": {
            "description": "令牌无效/过期"
          }
        },
        "summary": "刷新令牌：必须使用 Refresh Token（Bearer），换取新的 Access Token + 轮换 Refresh Token\n剩余寿命不足一半时重新签发 Refresh Token（滑动过期）；否则复用原 Refresh Token\n刷新令牌（Refresh Token 换新 Access Token + 轮换 Refresh Token）",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "默认",
      "url": "/"
    }
  ]
}